resource-aggregate: add basic GetEvents query
GetEvents query return list of events for given device and/or
resource that occured later than given timestamp. All parameters
(device, resource or timestamp) are optional. If some are not
set then no filtering based on that field will be done. If none
are set then all events for all devices are returned.