resource-aggregate: rework Delete call
Previously the implementation ran delete one-by-one
by deviceID and returned a list of device ids that had
at least one document deleted. However, this is not a
useful output because the caller cannot use it to determine
whether all documents with the deviceID were deleted,
or whether documents with given ID were not deleted because
no such documents existed before the call.
Instead of the output from the Delete command the db should
be queried again to get the current state.