Support unpublishing of resources by instanceIDs
Add instanceIDs field to UnpublishResourceLinksRequest to support
unpublishing of resources by both list of hrefs and/or list of
instanceIDs.

Use the new InstanceIDs field in coap-gateway handling of unpublish
request (which sends only instanceIDs values as part of the request
query) to avoid storing list of published resource hrefs. This
was necessary to translate received instanceIDs to hrefs used in
UnpublishResourceLinksRequest sent to resource-aggregate. Now the
list of instanceIDs can be send directly.