Add TestRequestHandlerDeleteResourceAfterUnpublish test case
This test case verifies a fix of a iotivity-lite bug during
unpublishing of resource links. Previously, resource links were
stored in a list and were deleted in the response handler of
the unpublish request when OC_STATUS_DELETE code was returned.
When some other code was returned the link remained in the list
of links to be deleted and caused unexpected behavior.