bundle: allow bundling to utilize setup from the previous run
In cases where the a config file does not exist, the bundling
process will attempt to generate it. All generated files are
stored within the "/data" directory. By utilizing a volume on
this directory, you can reuse the generated files for
subsequent runs. If you wish to regenerate all files, you can
set the environment variable "OVERRIDE_FILES" to "true".