coap-gateway: allow to use device redirect uri for signup
For a confidential OAuth client (e.g., server-side application),
the client application must protect the client ID and client secret,
which are used during the token request to authenticate the client
to the authorization server. For redirect URI it is optional.

In summary, the redirect URI itself is not a secret parameter in
the authorization code flow so it can be propagated to the device
and used in register action to the cloud.

This allows to have one provider in the cloud. Otherwise you need
to have provider for each redirecturi.