Associating Hub with Events for Enhanced Context (#1031)
Is is enhancing the context of events by associating them with their corresponding Hub(HubID). By establishing this linkage between the HubID and events generated by the hub (including those related to devices), we significantly improve the clarity regarding the source of each event. Through the incorporation of the HubID, clients gain the ability to precisely pinpoint the hub responsible for initiating a particular event. This alignment facilitates streamlined communication, ensuring that commands are directed to the correct hub where the device is connected.