certificate-authority: allow sets external caPool to signer for k8s
The Certificate Authority (signer.certFile) expects a complete
chain of certificates, starting from the intermediate sign cert up
to the root cert. However, in Kubernetes (k8s), the root cert is
not included in the tls.crt file; it is stored separately as
ca.crt. Therefore, this commit allows you to specify the CA file
as well via the signer.caPool. This change enables the resolution
of the full certificate chain, which is used during certificate
signing.