Service Monitoring: Establish Association between Service and Device (#1058)
* Service Monitoring: Establish Association between Service and Device

When a service has been terminated, and devices have not transitioned
to an "offline" status, they are incorrectly reported as "online." By
introducing this association, all devices belonging to the service
can now be updated to the "offline" state automatically.

This association is stored in the database, similar to device
resources. In this case, the hub  represents a device, and the content
is a special resource containing all services, each with a validity
period. This means that each service must update its validity into
the future. If a service goes down and its validity expires, any other
 service that updates its validity will trigger a check to see if any
service is expired. If any service is indeed expired, all associated
devices will also be updated to the "offline" state.