Extend the gRPC/HTTP Gateway API to support ETag in the resourceIdFilter.
This update enables the API to handle ETag values in the resourceIdFilter.
When using gRPC or HTTP streaming, the response for an unmodified resource
will have `data.status` set to `NOT_MODIFIED`, and `data.content.data`
will be empty. Additionally, for HTTP GET requests to
`/api/v1/devices/{deviceId}/resources/{resourceHref}` with an ETag in
the header, the response will return an HTTP status code indicating that
the resource has not been modified.