Configs: Add support for the data URI scheme property.
In certain cases, it is necessary to store property values as raw
data. This is applicable to certFile, keyFile, caPool, and
clientSecretFile, which now also support the data URI scheme. In
most instances, the data must be encoded using standard Base64 and
stored in the data scheme format: `data:;base64,<BASE64_DATA>`.