Test: fix TestCoAPGatewayServiceHeartbeat
This test should be constrained to a 3-minute time limit.
The reason for this limit is that the device's DTLS system identifies
a connection loss within 20 seconds for the initial ping, and subsequently,
there are 5 more pings, each separated by 4 seconds, with a timeout of 4 seconds.
Therefore, the total time for this sequence is 20 + 4 + (5 * (4 + 4)) = 64 seconds,
plus an additional 10 seconds for waiting for the device to come online.