Introduce test case for batch resource unpublish
This commit adds a new test case focused on the batch unpublishing
of resources. When resources are deleted, the CoAP gateway should
receive unpublish messages from the device, each containing the
instanceIDs of the deleted resources within the query. Given the
restricted size of CoAP packets, it's possible that the device may
need to emit multiple messages to transmit all instanceIDs.
This test case validates the functionality of the splitting
mechanism to ensure that all instanceIDs are accurately
transmitted.